requirements.txt
keras==3.9.2
numpy==2.1.3
opencv-python-headless==4.11.0.86
pillow==11.2.1
streamlit==1.45.1
tensorflow==2.19.0
